+++ src/EIDEProject.ts
@@ -2384,7 +2384,6 @@ class EIDEProject extends AbstractProject {
                     name: os.platform(),
                     includePath: <any>undefined,
                     defines: <any>undefined,
-                    compilerPath: "",
                     intelliSenseMode: "${default}",
                     configurationProvider: this.extensionId
                 };